fix(F23): split freshness closure and repair public dir gate
Why: public main expired 13 directory exceptions and blocked every pull request. This ports the already reviewed Cat Café #4165/#4167 repair instead of renewing a blanket bypass.

Moves the ten-file freshness closure cluster into a real submodule, updates 22 public test imports, and recounts the public-only exception set.

Source-Commits: cat-cafe e59df1f6ca, ba43220672

[小太阳·砚砚/GPT-5.6 Sol🐾]
Thread-Context: threadId=thread_mtgm45nd13bclepj catId=codex-sol
