+++ packages/api/src/domains/cats/services/freshness/closure/FreshnessClosureLegacyMigrationState.ts
