+++ packages/api/src/domains/cats/services/freshness/closure/FreshnessClosureStateMachine.ts
