+++ packages/api/src/domains/cats/services/freshness/closure/freshness-closure-telemetry.ts
@@ -3,7 +3,7 @@ import {
   freshnessClosureTransition,
   freshnessReplayFence,
   freshnessSuccessorPreflightCanceled,
-} from '../../../../infrastructure/telemetry/instruments.js';
+} from '../../../../../infrastructure/telemetry/instruments.js';
 
 export type FreshnessClosureTransitionKind = 'opened' | 'superseded' | 'committed' | 'blocked' | 'retried';
