+++ packages/api/src/domains/cats/services/freshness/closure/redis-freshness-closure-serialization.ts
