+++ packages/api/src/domains/cats/services/freshness/glass-box/FreshnessOutputRefiner.ts
@@ -4,7 +4,7 @@ import type {
   PublishedFreshnessAnnotation,
 } from '@cat-cafe/shared';
 import type { IMessageStore, StoredMessage } from '../../stores/ports/MessageStore.js';
-import type { FreshnessClosureStore } from '../FreshnessClosureStore.js';
+import type { FreshnessClosureStore } from '../closure/FreshnessClosureStore.js';
 import { recordFreshnessGlassBoxTransition } from './freshness-glass-box-telemetry.js';
 import {
   committedDecision,
