+++ packages/api/test/f254-freshness-closure-startup-recovery.test.js
@@ -2,10 +2,10 @@ import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 
 const { InMemoryFreshnessClosureStore } = await import(
-  '../dist/domains/cats/services/freshness/FreshnessClosureStore.js'
+  '../dist/domains/cats/services/freshness/closure/FreshnessClosureStore.js'
 );
 const { reconcileFreshnessClosuresAtStartup } = await import(
-  '../dist/domains/cats/services/freshness/FreshnessClosureStartupReconciler.js'
+  '../dist/domains/cats/services/freshness/closure/FreshnessClosureStartupReconciler.js'
 );
 
 const now = 7_200_000;
